another attempt
Some of us feel the original writeup is too long, repetitious, and
poorly worded.

If we're going to say anything at all, it should be concise, as
straight-forward as possible, and not belabor any one point for too
long.

Otherwise, I would rather just delete this repository altogether.

That said, if this were my project, and my project alone, I might take a
different approach.  My personal value set has some congruence with
Scalaz's approach, but is not perfectly aligned.

However, I'm open to Scalaz existing by the terms that Scalaz was
founded, and have tried to keep the part of the original writeup that
matches my experience using the library and working with other people
using or contributing to it.